-- Silver --
When a new shopping list is created and there are items, then delete shopping list was called. I modified the create shopping list method to do the one thing that was common between the two methods and then broadcast the shopping list created event. This then corrected the issue where the delete modal was being shown on new.